Updated User Profile Template and Search Feature
Updated the user profile template to display follower and following counts with links to respective views.
Added a Tweet button linked to the post-create URL.
Implemented a search feature within a form using the CSRF token.
Refactored the 'Users You Might Like' section to display a list of users in the template.